core/packages/pykit-bench/pyproject.toml
[project]
name = "pykit-bench"
description = "Generic accuracy benchmarking framework"
version = "0.1.0"
readme = "README.md"
license = { text = "MIT" }
keywords = ["benchmark", "accuracy", "evaluation", "testing"]
requires-python = ">=3.13"
dependencies = [
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest-benchmark>=4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pykit_bench"]
